requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
openai>=1.0.0
lancedb>=0.4.0
pypdf>=5.0.0
python-pptx>=0.6.23
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
tiktoken>=0.5.0
stix2>=3.0.0
dashscope>=1.17.0

